requirements.txt
matplotlib==3.8.2
mlflow==2.8.1
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
plotly==5.18.0
pyyaml==6.0.1
scikit-learn==1.3.2
seaborn==0.13.0
shiny==0.6.0
tqdm==4.66.1
